[Update] rename Factory, removeOneFunc, [Delete] CardPart in Factory
