+++ PokerGame/PokerGame/Card.swift
@@ -14,6 +14,7 @@ struct Card:CustomStringConvertible{
     
     var suit:Suit
     var rank:Rank
+
     
     //카드의 모양(suit)를 나타내줄 Enum
     //카드의 모양은 총 네가지로 고정되어 있으므로 예외처리를 원활하게 하기 위해서 Enum으로 설정한다.
